requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0
jupyter==1.0.0
pmdarima==2.0.4
fastapi==0.111.0
uvicorn==0.30.1
joblib==1.4.2
pydantic==2.7.1
mcp-server==0.1.25
mcp==1.12.3
qdrant-client==1.8.1
python-dotenv==1.0.1
requests==2.32.3
openai==1.98.0
langchain==0.3.27
langchain_community==0.3.27